project/python_backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
pillow==8.3.2
numpy==1.21.2
torch==1.9.0
torchvision==0.10.0
opencv-python==4.5.3.56
python-multipart==0.0.5